+++ app/page.tsx
@@ -1,5 +1,5 @@
 // layer
-import { CardListCont } from '@/views/card-list';
+import { CardListCont } from '@/views/home/card-list';
 
 export default function Page() {
   return <CardListCont />;
