+++ views/detail/post-detail/index.ts
