+++ views/detail/post-detail/styles/AiSummaryPres.module.scss
