+++ views/detail/post-detail/types.ts
